Front

+++ models/service/Agency.php
@@ -111,11 +111,24 @@ class Agency extends \app\models\Agency
             if (array_key_exists($key, $savedAdvertisements)) {
                 $advertisement = $savedAdvertisements[$advertisementData['external_id']];
 
+                // update advertisement data (if there are any changes)
+                if ($advertisement->last_update_date != $advertisementData['last_update_date']) {
+                    $advertisement->last_update_date = $advertisementData['last_update_date'];
+                    $advertisement->save();
+                }
+
                 // update secondary rooms
                 foreach ($advertisement->secondaryRooms as $secondaryRoom) {
 
-                    // TODO
-                    // update secondary rooms data here
+                    // update secondary rooms data (if there are any changes)
+                    if ($secondaryRoom->price != $advertisementData['price']) {
+                        $secondaryRoom->price = $advertisementData['price'];
+                        $secondaryRoom->save();
+                    }
+                    if ($secondaryRoom->detail != $advertisementData['description']) {
+                        $secondaryRoom->detail = $advertisementData['description'];
+                        $secondaryRoom->save();
+                    }
 
                     // sync room images
                     $savedImagesURLs = array();
@@ -382,16 +395,18 @@ class Agency extends \app\models\Agency
 
         foreach ($adsToDeleteIdies as $deleteId) {
             $advertisementToDelete = SecondaryAdvertisement::findOne(['external_id' => $deleteId, 'creation_type' => 1, 'agency_id' => $this->id]);
-            foreach ($advertisementToDelete->statusLabels as $label) {
-                $advertisementToDelete->unlink('statusLabels', $label, true);
-            }
-            foreach ($advertisementToDelete->secondaryRooms as $room) {
-                foreach ($room->images as $image) {
-                    $image->delete();
+            if ($advertisementToDelete !== null) {
+                foreach ($advertisementToDelete->statusLabels as $label) {
+                    $advertisementToDelete->unlink('statusLabels', $label, true);
                 }
-                $room->delete();
+                foreach ($advertisementToDelete->secondaryRooms as $room) {
+                    foreach ($room->images as $image) {
+                        $image->delete();
+                    }
+                    $room->delete();
+                }
+                $advertisementToDelete->delete();
             }
-            $advertisementToDelete->delete();
         }
 
         return $advertisements;
